add body component for signup page
