+++ lib/Pages/Signup/components/body.dart
@@ -1,0 +1,41 @@
+import 'package:flutter/material.dart';
+
+class Body extends StatelessWidget {
+  final Widget child;
+
+  const Body({
+    Key key,
+    @required this.child,
+  }) : super(key: key);
+  @override
+  Widget build(BuildContext context) {
+    Size size = MediaQuery.of(context).size;
+    return Container(
+      height: size.height,
+      width: double.infinity,
+      // here i can use size.width but i used double.infinity becuase they both same as far as i know
+      child: Stack(
+        alignment: Alignment.center,
+        children: <Widget>[
+          Positioned(
+            top: 0,
+            left: 0,
+            child: Image.asset(
+              "assets/images/signup_top.png",
+              width: size.width * .35,
+            ),
+          ),
+          Positioned(
+            bottom: 0,
+            left: 0,
+            child: Image.asset(
+              "assets/images/main_bottom.png",
+              width: size.width * .25,
+            ),
+          ),
+          child,
+        ],
+      ),
+    );
+  }
+}
